feat: filter group functionality
Functionality for selecting multiple filters and for the "All" reset,
which can return an array of the selected filters.

Adds an aria-label to the existing label with additional explanation of
the filter functionality.

Includes some mockup working filtering on placeholder card elements that
will be replaced with dynamic ideas article content in future work.