fix: reduce shift of filter buttons on keyboard focus
On keyboard focus, inline padding is used to avoid the focus outline
being cut off on the filter buttons. But it was only doing it on the
first element, which resulting in things shifting back in a jarring
way when tabbing to the second element. This keeps the padding whenever
any are focus-visible, not just the first.

Also reduces inline start padding to only what is needed to account for
the focus outline size and the focus outline offset.