feat: filter group functionality and datastore (#81)
Filter group functionality:
Functionality for selecting multiple filters and for the "All" reset,
which can return an array of the selected filters. Along with 
accessibility improvements. Includes some mock functionality
for filtering the placeholder ideas cards.

Datastore for fetching and sharing data:
The data store is a simple way to fetch, store, and share JSON data
between components. If components are requesting the same data, they
will use the same cached data instead of repeating requests.