feat: search for mobile
Always show search field in the mobile menu. Styles the mobile search
and the results within the mobile menu.

Updates escape key handling so the search field can be cleared before
closing the menu. Improves escape key handling so search results are
also closed when keyboard focus is on a search result link.