feat: recent ideas - option to pull tags from metadata (#138)
Add an optional value for the Recent Ideas block of "UseMetadataTags", for it to pull in the tags value from the current article page. Used for the article prefooter to show related stories/ideas. Also excludes the current page from list of fetched articles, so the same article you are on does not show in the list of related articles.