fix: refactor dark mode setting to use body class
Fix an issue discovered after implementing the saving and loading of
the color scheme setting in local storage, where the page will flash
with the wrong colors if the loaded setting does not match the default
system preference.

Moving this to a body class that is set during buildThemeToggle and
loadEager works because the body is already hidden until after that
logic runs (see the `appear`` class). This removes the need for the
repeated styles in delayed-styles from #113 .