feat: use smaller description text for page header without byline
Use design requested smaller text size for the page header component's
description, as used on the Careers page. Applies to usage of page
header without an author byline; the larger heavier weight font still
applies to the description when it's followed by the author byline, as
used on the homepage.