feat: dynamic job listings
Change Job-List component to list jobs dynamically from the careers
endpoint. Uses pre-existing markup, with the addition of a block wrapper
around all of the groupings.

The user entered content for the component has changed to be one row for
each discipline name, rather than a heading and a manually entered list
of jobs.