fix: link for single job listing component
The "Job Listing" component in the pattern library was previously
creating a link href with the job ID that does not match the format
of the actual job post links. They are unique paths and not named by ID.
This updates the component to use the linked title as the link.